#18 UPDATE use CrypterScene in EncrypterScene
