+++ mastermind/src/view/EncrypterScene.py
@@ -3,7 +3,6 @@ Created on 29.07.2017
 
 @author: Mirco
 '''
-from view.IScene import IScene
 import pygame
 from util import Settings
 from view import ViewSettings
@@ -11,30 +10,24 @@ from view.ViewSettings import COLORS_TO_RGB
 from view.Stone import Stone
 from ViewSettings import BLOCK_SIZE
 from view.DecrypterScene import DecrypterScene
+from view.CrypterScene import CrypterScene
 
 
 
-class EncrypterScene(IScene):
+class EncrypterScene(CrypterScene):
     
-    def __init__(self, color_choices_stones, current_combination_stones, code_given_in_colors):
-        IScene.__init__(self)
-        self.color_choices_stones = color_choices_stones
-        self.current_combination_stones = current_combination_stones
-        self.code_given_in_colors = code_given_in_colors
-        self.selected_combination_stone = None
+    def __init__(self, color_choices_stones, code_given_in_colors):
+        CrypterScene.__init__(self, color_choices_stones, code_given_in_colors)
         
         if Settings.DEBUG_LEVEL >= 1:
             print('enter EncrypterScene')
             
     def render(self, surface):
-        surface.fill(COLORS_TO_RGB['white'])
-        self._draw_color_choices(surface, self.color_choices_stones)
+        CrypterScene.render(self, surface)
         self._draw_row_of_stones(surface, self.current_combination_stones)
         if self.code_given_in_colors.count('') == 0:
             self.draw_continue_text(surface)
 
-    def update(self):
-        pass
 
     def handle_events(self, events):
         for event in events:
@@ -79,23 +72,6 @@ class EncrypterScene(IScene):
             True, COLORS_TO_RGB['black'])
         surface.blit(text1, (120, 200))
         
-    def _draw_color_choices(self, surface, color_choices_stones):
-        if not color_choices_stones:
-            self.color_choices_stones = self._init_color_choices_stones(surface)
-        for stone in self.color_choices_stones:
-            stone.render(surface)
-            
-    def _init_color_choices_stones(self, surface):
-        counter = 1
-        color_choices_stones = []
-        for color in Settings.COLORS:
-            stone = Stone(COLORS_TO_RGB[color], BLOCK_SIZE, BLOCK_SIZE,
-                          surface.get_width() / (len(Settings.COLORS) + 1) * counter - BLOCK_SIZE / 2,
-                          surface.get_height() * 0.8)
-            color_choices_stones.append(stone)
-            counter += 1
-        return color_choices_stones
-    
     def _draw_row_of_stones(self, surface, current_row_of_sprites):
         if not self.current_combination_stones:
             self._init_current_row_of_stones(surface)
