+++ mastermind/src/view/TitleScreen.py
@@ -32,4 +32,4 @@ class TitleScene(IScene):
     def handle_events(self, events):
         for event in events:
             if event.type == pygame.KEYUP and event.key == pygame.K_RETURN:
-                self.manager.go_to(EncrypterScene([], [], []))
+                self.manager.go_to(EncrypterScene([], []))
